refactor(gateway): Q11.e — extract rest_v1/tasks/subscriptions.rs (GAR-635)
Move task-subscription section (plan 0079 / GAR-539, ~279 LOC) from
rest_v1/tasks/mod.rs into rest_v1/tasks/subscriptions.rs.

- New: tasks/subscriptions.rs (299 LOC) — 3 handlers + 2 structs
- mod.rs: 2341 → 2067 lines; add `pub mod subscriptions; pub use subscriptions::{...}`
- openapi.rs: update 3 handler paths to super::tasks::subscriptions::*

Zero behavior change — pure structural refactor (plan 0140).
