+++ crates/garraia-gateway/src/rest_v1/openapi.rs
@@ -137,9 +137,9 @@ impl Modify for SecurityAddon {
         super::tasks::labels::delete_task_label,
         super::tasks::labels::assign_task_label,
         super::tasks::labels::remove_task_label_from_task,
-        super::tasks::subscribe_to_task,
-        super::tasks::list_task_subscriptions,
-        super::tasks::unsubscribe_from_task,
+        super::tasks::subscriptions::subscribe_to_task,
+        super::tasks::subscriptions::list_task_subscriptions,
+        super::tasks::subscriptions::unsubscribe_from_task,
         super::tasks::move_task,
         super::tasks::list_subtasks,
         super::audit::list_audit,
